+++ MareSynchronos/PlayerData/Pairs/PairManager.cs
@@ -288,7 +288,7 @@ public sealed class PairManager : DisposableMediatorSubscriberBase
     }
 
     private Lazy<List<Pair>> DirectPairsLazy() => new(() => _allClientPairs.Select(k => k.Value)
-        .Where(k => k.UserPair.Groups.Contains(Constants.IndividualKeyword)).ToList());
+        .Where(k => k.UserPair.Groups.Contains(Constants.IndividualKeyword, StringComparer.Ordinal)).ToList());
 
     private void DisposePairs()
     {
